Making withLoggerNameFromLine public in core (#203)
