+++ core/js/src/main/scala/zio/logging/PlatformSpecificLogAppenderModifiers.scala
@@ -1,0 +1,3 @@
+package zio.logging
+
+trait PlatformSpecificLogAppenderModifiers {}
