+++ core/shared/src/main/scala/zio/logging/LogContext.scala
