+++ core/shared/src/main/scala/zio/logging/LogFormat.scala
