+++ core/shared/src/main/scala/zio/logging/LogWriter.scala
