+++ core/shared/src/main/scala/zio/logging/Logging.scala
